backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# AI Models
anthropic==0.39.0

# Google Calendar API
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# Database (PostgreSQL via Supabase)
supabase==2.9.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
pytz==2024.2

# Testing (optional for later)
pytest==8.3.3
pytest-asyncio==0.24.0
